# derive: Avoid emitting provided PartialEq, PartialOrd methods for c-like enums