@bluss 2016-02-29
Using perf stat -r5 to measure compile time on a [small test crate with 140 ten-variant enums (source)](https://gist.github.com/bluss/2e47edf1b062b8cdc019).

- Before: 2.54 seconds time elapsed   ( +-  9,03% )  2170,409698      task-clock (msec)
- After: 1.55 seconds time elapsed   ( +-  9,60% )  1305,038461      task-clock (msec)