@bluss 2016-02-29
Timing on src/libgraphviz (which has two enums that match this optimization)

libgraphviz debug build (perf stat -r20)

- before 0,90 seconds time elapsed ( +-  3,18% )
- after 0,76 seconds time elapsed   ( +-  2,67% )

libgraphviz -O build  (perf stat -r20)

- before 1,72 seconds time elapsed  ( +-  2,86% )
- after 1,46 seconds time elapsed  ( +-  4,07% )

as a control, librand (no enums match) debug build

- before 1,53 seconds time elapsed  ( +-  4,48% ) 
- after 1,49 seconds time elapsed   ( +-  3,64% )