@brson 2016-02-29
r=me

Great little optimization.