@durka 2016-02-29
Just for paranoia, did you confirm no runtime impact?