@bluss 2016-03-01
I did some tests using only the `partial_cmp` for a c-like enum versus the full ensemble `partial_cmp, lt, le, gt, ge`. It seems like there is no performance change.

So I added a change to treat `PartialOrd` the same way, it has four methods we can let have just the default implementation. It saves a lot of code for an enum of many variants. However, `PartialOrd` is not at all as common as `PartialEq`.

Since the code is general for "enum or struct has no fields", it also affects unit structs, which are relatively common.